move container to home
